requirements.txt
# بعد از تست لوکال نسخه‌ها را pin کن با pip freeze
telethon
pyrogram
requests
aiohttp
python-dotenv
